+++ ui/src/components/dataset/ingestion/FileListAutoComplete.vue
@@ -9,7 +9,7 @@
     @clear="emit('clear')"
     @select="
       (file) => {
-        console.log('@select: selected file:', file);
+        // console.log('@select: selected file:', file);
         onFileSelect(file);
       }
     "
@@ -27,7 +27,6 @@
 </template>
 
 <script setup>
-
 const props = defineProps({
   disabled: { type: Boolean, default: false },
   basePath: { type: String },
@@ -48,53 +47,25 @@ const emit = defineEmits([
 
 const searchText = computed({
   get: () => {
-    console.log("getting searchText,", props.searchText);
     return props.searchText;
   },
   set: (value) => {
-    // searchText.value = value;
-    // const _searchText = +value.path;
-
-
-    console.log("emitting searchText,", value);
-    console.dir(value, { depth: null });
-    console.log(typeof value);
-    console.log('----')
-
     emit("update:searchText", value);
   },
 });
 
-const fileList = ref([]);
-const loading = ref(false);
 // const searchText = ref("");
 const basePath = computed(() => {
   return props.basePath.endsWith("/") ? props.basePath : props.basePath + "/";
 });
 
 const onFileSelect = (file) => {
-  console.log("onFileSelect:", file);
-  console.dir(file, { depth: null });
-  // const _searchText = basePath.value + value.path;
-  emit("update:searchText", file.path.slice(file.path.indexOf(basePath.value) + basePath.value.length));
-  console.log("update:searchText", file.path.slice(file.path.indexOf(basePath.value) + basePath.value.length))
-
-  // searchText.value = file.path.slice(
-  //   file.path.indexOf(basePath.value) + basePath.value.length,
-  // );
+  emit(
+    "update:searchText",
+    file.path.slice(file.path.indexOf(basePath.value) + basePath.value.length),
+  );
   emit("update:selected", file);
 };
-
-// const setRetrievedFiles = (files) => {
-//   fileList.value = files;
-//   console.log("retrieved file list");
-//   console.log(fileList.value);
-//   emit("filesRetrieved", fileList.value);
-// };
-
-// const selectFile = (file) => {
-//   console.log("Selected file:", file);
-// };
 </script>
 
 <style lang="scss" scoped>
